updating the chapter card to match brand new design
