+++ src/components/card.stories.js
@@ -50,7 +50,6 @@ storiesOf('Card', module)
   .add('Catalog Item Cover', () => (
     <Card style={styles.card} shadowStyle={BOX_STYLE}>
       <CatalogItem
-        isCourse
         title="Predicting the future"
         subtitle="Coorpacademy"
         progression={progression}
@@ -72,7 +71,6 @@ storiesOf('Card', module)
   .add('Catalog Item Card', () => (
     <Card style={styles.card} shadowStyle={BOX_STYLE}>
       <CatalogItem
-        isCourse
         title="Predicting the future"
         subtitle="Coorpacademy"
         progression={progression}
