+++ src/components/catalog-item.stories.js
@@ -5,7 +5,7 @@ import {storiesOf} from '@storybook/react-native';
 
 import image from '../__fixtures__/assets/landscape-1.jpg';
 import type {Progression} from '../types';
-import {AUTHOR_TYPE} from '../const';
+import {CARD_DISPLAY_MODE, AUTHOR_TYPE} from '../const';
 import {handleFakePress} from '../utils/tests';
 import translations from '../translations';
 import CatalogItem from './catalog-item';
@@ -16,9 +16,8 @@ const progression: Progression = {
 };
 
 storiesOf('Catalog Item', module)
-  .add('Learner', () => (
+  .add('Adaptive', () => (
     <CatalogItem
-      isCourse
       title="Predicting the future"
       subtitle="Coorpacademy"
       progression={progression}
@@ -33,9 +32,8 @@ storiesOf('Catalog Item', module)
       section="finishLearning"
     />
   ))
-  .add('Microlearning', () => (
+  .add('New', () => (
     <CatalogItem
-      isCourse={false}
       title="Predicting the future"
       subtitle="Coorpacademy"
       progression={progression}
@@ -48,6 +46,60 @@ storiesOf('Catalog Item', module)
       testID="catalog2"
       universalRef="foobar"
       type="learner"
+      section="finishLearning"
+    />
+  ))
+  .add('Adaptive/New/Certified/Coorp', () => (
+    <CatalogItem
+      title="Predicting the future"
+      subtitle="Coorpacademy"
+      progression={progression}
+      image={image}
+      authorType={AUTHOR_TYPE.COORP}
+      badge={translations.new}
+      isAdaptive
+      displayMode={CARD_DISPLAY_MODE.COVER}
+      isCertified
+      onPress={handleFakePress}
+      testID="catalog3"
+      universalRef="foobar"
+      type="learner"
+      section="finishLearning"
+    />
+  ))
+  .add('Adaptive/New/Certified/Custom', () => (
+    <CatalogItem
+      title="Predicting the future"
+      subtitle="Coorpacademy"
+      progression={progression}
+      image={image}
+      authorType={AUTHOR_TYPE.VERIFIED}
+      badge={translations.new}
+      isAdaptive
+      displayMode={CARD_DISPLAY_MODE.CARD}
+      isCertified
+      onPress={handleFakePress}
+      testID="catalog4"
+      universalRef="foobar"
+      type="learner"
+      section="finishLearning"
+    />
+  ))
+  .add('Microlearning', () => (
+    <CatalogItem
+      title="Predicting the future"
+      subtitle="Coorpacademy"
+      progression={progression}
+      image={image}
+      authorType={AUTHOR_TYPE.VERIFIED}
+      badge={translations.new}
+      isAdaptive
+      displayMode={CARD_DISPLAY_MODE.CARD}
+      isCertified
+      onPress={handleFakePress}
+      testID="catalog4"
+      universalRef="foobar"
+      type="microlearning"
       section="finishLearning"
     />
   ));
