+++ src/components/catalog.js
@@ -72,8 +72,6 @@ export const getAuthorName = (card: DisciplineCard | ChapterCard): string | void
 class Catalog extends React.PureComponent<Props> {
   props: Props;
 
-  static isCourse = (type: string) => type === 'course';
-
   handlePress = (item: DisciplineCard | ChapterCard) => () => this.props.onPress(item);
 
   handleLogoLongPress = () => this.props.onLogoLongPress && this.props.onLogoLongPress();
@@ -103,7 +101,6 @@ class Catalog extends React.PureComponent<Props> {
 
               <Card style={styles.card} shadowStyle={BOX_STYLE}>
                 <CatalogItem
-                  isCourse={Catalog.isCourse(cover.type)}
                   title={cover.title}
                   subtitle={cover.authors.map(author => author.label).join(', ')}
                   progression={{
@@ -142,7 +139,6 @@ class Catalog extends React.PureComponent<Props> {
                     <View style={styles.cards}>
                       <Card style={styles.card} shadowStyle={BOX_STYLE}>
                         <CatalogItem
-                          isCourse={Catalog.isCourse(item.type)}
                           title={item.title}
                           subtitle={item.authors.map(author => author.label).join(', ')}
                           progression={{
@@ -173,7 +169,6 @@ class Catalog extends React.PureComponent<Props> {
                       {nextItem && (
                         <Card style={styles.card} shadowStyle={BOX_STYLE}>
                           <CatalogItem
-                            isCourse={Catalog.isCourse(item.type)}
                             title={nextItem.title}
                             subtitle={nextItem.authors.map(author => author.label).join(', ')}
                             progression={{
