+++ src/components/level-end.js
@@ -156,7 +156,6 @@ const styles = StyleSheet.create({
 const {width: screenWidth} = Dimensions.get('window');
 
 class LevelEnd extends React.PureComponent<Props> {
-  static isCourse = (type: string) => type === 'course';
 
   handleCardPress = (item: DisciplineCard | ChapterCard) => () => this.props.onCardPress(item);
 
@@ -246,7 +245,6 @@ class LevelEnd extends React.PureComponent<Props> {
                       <Text style={styles.title}>{translations.relatedSubjects}</Text>
                       <Card type={CARD_LAYOUT.CONTAIN} style={styles.card} shadowStyle={BOX_STYLE}>
                         <CatalogItem
-                          isCourse={LevelEnd.isCourse(recommendation.type)}
                           title={recommendation.title}
                           subtitle={recommendation.authors.map(author => author.label).join(', ')}
                           progression={{
